Add key path support for methods on `Mockable`
This allows the client of a mock to type less code in order to use
`registerResult(for:result:)`, `calls(to:)`, `resetCalls(to:)`, `hasCalled(_:)`,
and `hasCalled(_:numberOfTimes:)`. Instead of duplicating the name of the
`Mockable` instance like this:

someMockableInstance.calls(to: someMockableInstance.someMockedMethodRef)

it's now possible to instead use a key path like this:

someMockableInstance.calls(to: \.someMockedMethodRef)

This also means that clients can't accidentally pass a ref from mock A to
a method on mock B, but I'm not sure if that has ever been a problem in
practice.